+++ tests/testthat/test-gao_links.R
@@ -1,4 +1,5 @@
-test_that("gao_links() returns character vector", {
+test_that("gao_links() returns a data.frame with correct columns", {
   links <- gao_links()
-  expect_type(links, "character")
+  expect_s3_class(links, "data.frame")
+  expect_named(links, c("url", "title", "report_id", "published", "released", "summary"))
 })